First step to an authentication system that uses iChain.
This commit adds a model unregisteredperson to the webclient
which is used to transport the registration wish to the 
frontend.
The privacy_controller is a NOP controller that only checks
for the X-username key in the HTTP header that ichain sends.
The ichain login mask only comes up if this controller is 
accessed.
User_controller got new actions register and request_ichain.
application controller checks for the ichain subsystems enabled
through the config and extracts the user info from the header
-> still TODO: both authentication systems usable in parallel