Give the obs package a home in the svn.
Also added a scripte 'distribute' which does the following steps:

  checkout      get the current code from the build server
  freshtar      update the tar ball
  diff          compare what is in SCM with what is in obs
  update        update the obs workign directory with what is in the SCM
  localbuild    build it on the local machine
  upload        upload to the build server

It is somewhat related to oscupstream, which however focusses on
creating a tar ball and uploading that, which is "freshtar" and
"upload" only.

The above is a working cycle for someone who _manually_ maintains a
package.  the distribute script could possibly use oscupload at some
point.

The upload at this time is not fully implemented yet as it requires
removing the old tar ball and adding the new one.