rework the way the log file is loaded into the live_build_log
formerly the full log file was put into the initial HTML file,
taking forever. Now that the backend correctly implements end,
we can limit the chunks to 64K, so we can continuely load 64K
chunks into the web page and the user has a way to stop it and
look only at the beginning. 

Technically we would also be able to provide a "skip to the
end" feature, but so far I want to have more feedback on the 
current way