Add LDAP as an authentication method for the API.
Attempts to authenticate user and password against LDAP servers.  If
LDAP authentication fails, then the OBS database is used.  This allows
test users to be entered into the OBS database instead of having to
add them to the LDAP servers, which may be not be allowed due to
corporate policies.

Adds a install requirement on 'net/ldap'.
LDAP usage can be configured with the following new options

LDAP_SERVERS - a ':' separated list of ldap server hostnames
LDAP_SEARCH_BASE - search base for finding LDAP users
LDAP_SEARCH_ATTR - attribute to use as login name for OBS.  Defaults to
                   sAMAccountName
MAX_LDAP_ATTEMPTS - max number of attempts to try to connect to servers