[api] implement two new commands for packages
 POST /source/<project>/<package>?cmd=disable&repository=:opt&arch=:opt&flag=flag
 POST /source/<project>/<package>?cmd=enable&repository=:opt&arch=:opt&flag=flag

For now this is only for packages as I think the flags handling in
the database needs a bigger change - and I want to verify the http
interface. And it's not clear to me what should happen if you call disable
twice or disable+enable. My thinking is to write a cleanup function that
will only keep the latest state