[api] change the backend test suite logic
move the test backend logic into an extra script that
can easily be started from application controller. This way
the unit tests can continue playing only with the database
and do not need to care for the backend

This makes it much more reliable to do test coverage, so
adding little rake task. You can check with rake test:coverage,
that we're testing less than 50% of our code ;(