getbuildcode: serve worker qemu-cross images per build/qemu-reg
When the worker starts running non-native code using binfmt_misc
registration, it is best if it uses the server's static qemu images in
its virtual machine. This makes worker operation more portable and
its build behavior more consistent.

To identify which of the many qemu-* images make sense to ship to the
worker, look up the names of images from from the build/qemu-reg
interpreter column, and shorten the list when we find we do not have a
matching local image.

The fallback on the worker side is to use the worker machine's qemu
image, so it's good if this machine provides them, and no big deal if it
does not.

There is a coordinating change to build/init_buildsystem.

Signed-off-by: James Perkins <james@loowit.net>