Steps regarding a sane source documentation.
The files README and INSTALL where added/expanded but not yet finished.
It is planned to provide initial documentation to survive the first
steps after checking out the project. The docs try to be as terse as
possible and relegate to the openSUSE wiki where appropriate. It is
planned to out-phase all the docs that are spread across the project
tree. Also, the script 'git-update' was moved from 'src/webui/scripts'
to 'dist', where other useful scripts are hanging around.