[api] Add 'enabled' flag to architectures table.
Currently, the table also has a 'selectable' flag that is currently
unused. Both will replace the custom code of webui/osc for determining
which archs should be displayed to the user (selectable) and which are
provided by the backend (enabled).

In the future, the backend may even update the database table when
schedulers come and go. Or, in an ideal world, the backend could
start/stop/enable/disable schedulers based on the arch table's 'enabled'
column. An administrator could then set this in an admin interface
instead of having to interface with the schedulers directly.