[api] Make seed.rb invokable multiple times again.
Seed.rb is for default production data. Additional data may be added
over time when new migrations appear. Previously, seed.rb was not able
to be run again as it didn't check if the database entries it creates
are already there. Additionally, it now displays some informational
status messages on what is seeded.