[webui] Cache and combine JS files that are always sent.
One bigger file is better than five smaller ones, less connections and
faster download. The file is re-generated by Rails if the input JS files
change. Note that the resulting JS file is not 'minimized' in any way,
it's just a concatenation of the other files.

Note: This would work for CSS as well but we can't combine the three CSS
      files of application.html as they provide alternate layouts.