[api] Also allow PUT'ing raw XML to /configuration URL.
This is only to be consistent with the existing API controllers. In
fact, this is bad behavior as raw XML as PUT data is anything but
x-www-form-urlencoded data. Thus Rails fails to automagically parse this
into Ruby structures.

The configuration controller solves this by first trying to interpret it
as such and fall back to XML parsing if it was unsuccessful. We think
about adopting such behavior for the other controllers.