[webui] Fix view_file links for deleted files in diffs.
Deleted files aren't available in the diff source but in the diff
target. The 'sourcediff' view currently doesn't take the target as
parameter, thus the deleted file is taken from source_revision - 1. This
is a bit hacky but works. A saner solution would be to pass the diff
target revision to the 'sourcediff' partial but that would complicate
the 'render :partial => ...' line quite a bit without too much win.