[webui] Fix initial request display for projects/packages.
The respective views used '@requests' for as the request list, this is
wrong, it's the users request list. Removing that default parameter
leads to the partial fetching the correct list via Ajax (like it would
do if the user selects a different filter). If one wants to initially
display requests, one would have to provide equal default parameters.

Also fixes bnc#719508