[backend] copy all known tree meta data
Commit 97c1d6c61 introduce only a partial copy of the required meta
data. This commit copies additionally to the srcmd5 metadata, also
the meta data of: lsrcmd5, xservicemd5, lservicemd5 - if available.

This is required for linkproject=all projects to avoid redudant source
service triggers. As the meta data did not get copied, handleservice
always triggered a source service run for linkedprojects. Even when the
sources already got copied with copyfiles().