[webui] use a hash instead of the full activexml in project status
As it's a very complex algorithm working on the big XML sent from the API
this really pays out. The initial conversion from XML to HASH takes 6 seconds
(it's 3.8MB of XML and it's ruby :), but the algorithm not using XML but native
ruby leads to performance improvement.

benchmark: querying factory status for 4 users of various levels of factory involvement.

before: user - 109.34s
after:  user -  34.62s

speedup (with very little code changes): 3x