[backend] the file is now named $package-appdata.xml
I would write more about this change, but Michael always insists
that I make sure the commit log is shorter than the actual code change,
so I'll try to limit myself to the bare minimum to describe the change.

Of course the code itself is clear and the change is even more clear,
but a short sentense should be allowed.

Oh, sorry I again broke the rule.