[api,backend] move request notifications to generic route in backend called from API. the backend no longer knows about requests
I left the code in the backend for now to allow master backends to run
behind 2.3 apis, but this is going to be removed for the beta.

One possible problem is the multi action support. I didn't really test
it, but it sounds possible that the generic route is not prepared to
have multiple actions