[api] merge Package and DbPackage
The bssql plugin in activexml is making things very complex
and hard to understand. I bet it was a clever idea at some
point, but the whole activexml is too complex for its own
good, so get rid of the seperation of "Package as XML" and
"Package in database" and just use xmlhash where we need
the former.