[activexml] make activexml more robust with capybara
The problem is that capybara runs the webui in a thread and selenium in
another. This leads to the problem that our libraries are tested for
thread safety - which they are not. So the work around is adding a
mutex, but this only makes sure http_do calls happen in one thread only.

But the real fix is a change in the API as everything accesses the
global transport object and sets headers and then assumes the next
http_do call will use that headers - which is only true if there is only
one thread. But having multiple threads in the test suite sounds like
something we want to have