[api] validate package name to be valid
the backend does only allow specific names and we don't want to notice
it when saving to the backend, but before that

In addition to that also make sure to run most things that put something
into the backend in a transaction. Having things in the database the
backend refused, can hurt us badly.