[backend] Avoid endless rebuilds if qemu-linux-user is missing
A common error happening on the build.opensuse.org instance
is that buidl is enabled for armv7l, but no or not the right
qemu-linux-user is installed. In those cases the builds
are looping endlessly on the build cluster, wasting a lot of
CPU. Detect that case and end it.