bs_admin: add feature to clean up sources
Delete sources older than x days but keeps at least n revisions, both
given as parameter.

Deletes sources, keeps revisions if an other revision links to it,
if this revision will be kept (because it is younger than x days or
in the count of revisions that has to be kept).
Rewrites .rev files for the projects.