[backend] Clean up emulator build invocation code
For the emulator case, a lot of the options were not passed
down to the build invocation, like for example the filesystem
to create, which defaulted to the site-wide filesystem instead
of the worker-defined filesystem type.

The one two differences we want to have is to disable statistics
(as they slow down the emulated build) and use a different way
to invoke the emulator mode, otherwise the rest should be passed
down as is.