[webui] Rename Home to Watchlist
Description of watchlist is also misleading

The thing shows only the watchlist nothing else, so there are no
involved projects, plus you can watch something even if you are not
involved.

Make the link empty for watchlist

Actually having no js is not working for obs at all so no need to
have working link here.
This actually solves the issue that javascript did watchlist, but
the nojs is pointing to the home page.