[api] replace the average check with a black list
the average is not good enough as measure as removing a one line
function increases the average, which is not what we want to prevent
per se.

Now making functions more complicated than 40 points deserves an
entry in the black list or a thought how to write it differently.

Note that codeclimate already gives a B for one 35 point function :)