[api] disallow the creation of new group requests
request actions of type group were official never supported
but there is build.opensuse.org which contains quite some of these
requests. However, it is not used there anymore, so dis-allow to create
new requests. But we do validate that the code is still working.
FIXME3.0: drop this code and drop these actions from database.