bs_publish: RPM: Make writing the project public key and referencing it in the .repo file consistent.
This commit fixes the following bug/situation:
- a global keyfile is defined (option "keyfile"), but no project-specific sign key exists
- the RPM repository will be published with the public key from the global keyfile
- the .repo does not reference this public key file with gpgkey= because the code only triggers if there is a project-specific sign key. Instead, for gpgkey= the "gpg_standard_key" config value will be written to the .repo file ("gpg_standard_key" usually points to a local file on the OBS host)

In the end, adding the repository using the .repo file will most likely fail because gpgkey= now contains a path to a local file which doesn't exist on the machine where the repository should be added. The solution is to reference the project's public key file when it was generated.

This commit also introduces two behavior changes:
- "gpgcheck=1" won't be added to the .repo file if signing was not configured
- when signing was configured, but "gpg_standard_key" was not configured, the .repo file did contain "gpgcheck=1", but not "gpgkey=". Now the publisher will fail.