[api] adapt path elements on branching with "add_repositories"
This is default for the maintenance case.

It checks if used repositories from the branch targets are depending on each other.
In this case the repositories within the branch project must also depend on each
other. So the path elements get adapted.

This is the typical case when two packages are built in one maintenance incident
for the same project and they live in different service packs. If one package
is using the other one at build time the repositories within the branch
project must depend on each other directly to get the same build result.