[backend] refactor service error handling
Unify with link error handling: the last line in the
error file is the error summary, which is used when
reporting the errors. The complete file is available
via the _serviceerror file in the xservicemd5.

Unfortunately we need to do this a bit hackish, as
the _serviceerror is stored in the tree dir.