[webui] Update OBS redirect after login
After login OBS users get redirected to the page they initially visited. So far
this was done via hidden fields in the views and parameters that were processed
in the controller.
An attacker could use those parameters to redirect to an untrusted side.
This commit stores the last visited page in the session store to avoid that kind
of attack.

Conflicts:
	src/api/app/controllers/webui/user_controller.rb
	src/api/app/controllers/webui/webui_controller.rb
	src/api/test/functional/webui/signup_test.rb