[backend] rework 100-continue handling
We now postpone contunue handling to when the body needs
to be read in. This allows us to send errors without needing
to read the body.
We also read and discard the body when doing a reply, so that
clients do not get a SIGPIPE when trying to send the body.