[api] Optimize default_scope for packages
Previously we selected all packages which are not included in a given range (forbidden projects) which caused performance issues.
This commit goes now the other way around and removes the forbidden projects from all projects.
This is a performance tweak which assumes that there are more allowed projects than forbidden ones.