[api] Include user's watchlist when admin requests user data (issue#1204)
Other users are not supposed to see a user's projects watchlist. Though if an
admin requests user data via the API (/person/$user), he should be able to see it.
Otherwise there is a risk that an admin unintenionally deletes this data, eg. when
editing user data via osc (see issue#1204).