[ci] Update save_meta test
Previous test was comparing the resulting strings, while the new tests are
verifying the actual xml. Disadvantage of previous test code was that it failed
when running tests on a SLE11 SP3 system. This most likely happened due to changes
in capybara's html parser (the xml we are testing comes from an html form).