[webui] Move request permission check in lock statement
User reported that it was possible to accept a request several times.
Changing the request state is already wrapped in a transaction but
checking the permissions and creating the history is not. So finally
moved permission check and creating history also in the with_lock
block to avoid race conditions. This will fix #1157.