[backend] touch dirty flag in incident local dependening repositories
This is a honest attempt to solve bnc#952404, possible release of
not matching release numbers. The api is blocking a release when
the repos are not in clean state. But under high load the dirty
flags may not be set early enough after a package got build.

However, this can only happen under high load as intermediate
state. So it is hard to test this atm.