[api][webui] avoid wrong usage of User.can_modify_project?
can_modify_project? must not be used on non-stored projects. We
check only at store time creation permissions, so we would just skip them here.

There is a convenience method in Project.check_write_access(!) to
call the right check depending on if a project got created already or not.

Raise an error, when this is not done right. Fixes obs#1291