[webui] Prevent package submit requests from failing because of trailing whitespaces, issue#1248
When a user accidentally adds a trailing or leading whitespace in the targetpackage field,
OBS was running into an exception during creation of the submit request. This commit causes
OBS to remove trailing and leading whitespaces before passing the data to the BsRequest class.

Ideally there would be some javascript to prevent such user in put, but we currently
don't have that.