[api] Force login for all API calls
Except /public, /trigger and /about route. This is necessary to be consistent with our public instance which
uses the iChain login proxy and forces also login for all request except the described ones.
Furthermore this caused a bug in the project list because OSC always tries to access a route without
user credentials and only sends the credentials when the route returns 401. This was not the case for
the project list and it was not possible to receive for instance hidden projects.
Fix #833