[webui] Show all open requests of a project
instead of only user related requests. This caused some misunderstandings recently because the project page
only showed open requests when thei're related to the user and not all open requests. Now we show all open
requests of a project. Furthermore this commit does some refactoring of the according function.
This will fix #1232.