[webui] redirect to home after login from /user/do_login
If a user enters wrong credentials in login widget and he gets redirected to /user/do_login.
If login happens with correct credentials now the user is logged in but stays on the page and sees the login form again in master, or gets an 404 in 2.6.

Now he gets redirected into home if the user comes from the login page.

fixes #1557