[doc] Update rspec readme
Remove adding skip to tests as a valid approach to handle untested code. This
was leading developers to the wrong impression that adding skips in our test suite
is a valid approach when migrating tests, or even writing new features.

It's not! Skips should be used in rare exceptions, when we can't fix a broken
test right away.
In any other case we should aim for writing the test!