[backend] Fix addding an interconnect if using partioning
With using multiple repservers as partitions adding a interconnect may not
work properly, if the first use of the interconnect happens on a different
partion.
In this case the scheduler states an error 'bad config' and need a
cold start to resolve this situation.
The reason is that the %remoteprojs map on the local scheduler has no entry
for this project in addrepo().
So if a remote project was added all repservers need to be notified about the
remote project.