[backend] avoid endless builds with broken base system
We do not try more then 5 times to build a job that fails with "badhost"

badhost usually means that the build worker thinks he is the problem,
because he can not even start the guest VM kernel.

But it can also be caused by the projects, for example by providing
a broken kernel, initrd, glibc or alike.

extend jobstatus to log the number of attempts for a job