[backend] redo badhost limitation code
The old commit didn't work with masterdispatched jobs and also
did not clean up the attempts hash. We now handle the badhost
limit with the standard badhost hash.

This means that we have to transport the job in the badhost event,
so we can no longer mis-use the job field for the worker name. So
we now have a 'worker' element in the event as well.