[webui] Ensure numbers in session[:request_numbers] are unique
session[:request_numbers] are used for allowing users to go through
a number of requests belonging to a project.
Since the implementation of selecting the next or previous item of
that list is using it's index, having duplicates causes a loop which
ignores requests that are not enclosed by the duplicates.
Eg. [10, 5, 10, 3] would never reach 3.