[webui] don't confuse a user with "maintenance" project
The check box is only confusing the user in first place.
It is not a project for preparing a maintenance update (but for creating
a new maintenance team coordination place). This is so rarly used
that it harms more then it helps here.
(if we re-implement, we should do it via a selection box for all the
 other (and more often used) types)