[ci] Silence warnings when running lib/tasks/test_webui.rake rake tasks
That's already the case for current OBS (running on rails 4.2 and rake v10.4),
and needs to be explicitly disabled with upgrading OBS to rails 5.

This caused our logs in travis to be spammed with warning messages,
like "warning: loading in progress, circular require considered harmful", after
switching to rails 5.
That circular require issue should be fixed nonetheless, eventually.