[ci] Minimize usage of raw_put in our tests
Before rails 4 it was necessary to use raw_put to send raw put data
to a controller.
Since rails 4 post is checking the second parameter and treats it as
raw put data, if it's a string.

As a side effect we also stop using the customized raw_put from our
test_helper.rb