bs_repserver: Fix different xml results return the same md5sum.
The resultlist node contains an entry 'state' which holds the calculated
md5sum for the corresponding message.

The 'code' and 'state' of the result nodes have not been taken into account
when calculating the md5sum for resultlist 'state'. Thus a client watching
those entries may wait forever, e.g. using ``osc.core.get_package_results``
which uses the oldstate HTTP-GET parameter. Using this parameter the server
will not send any new responses until the md5sum has changed.
With this patch the corresponding 'code' and 'state' entries will be
appended to the $state variable which will be used for the md5_hex calculation.