[api][webui] Set BsRequestCounter correctly
currently it is the number of the last BsRequest. However, it should be the next available number.
This causes ActiveRecord::RecordNotUnique when creating a new BsRequest.

- Wrap database actions into transaction
- Remove redundant initialisation code - this is done now in 20160824132643_fix_bs_request_counter.rb
- Introduce BsRequestCounter model to use ActiveRecord methods
- Remove redundant before_create call, assign number will return in the second call immediately

As assign_number is called in a before_save filter so it is still possible that save fails and the BsRequestCounter is already increased.

Fixes #2068.