[webui] Add note for project maintainers
Sometimes, project maintainers overstep package maintainers by accepting
submit requests, without giving the target package maintainers a chance
to review the request. This commit adds a note to submit requests, if
the user (who is reviewing the request) is a project maintainer but not
a package maintainer.

Related to issue openSUSE/open-build-service#1970