[backend] no longer fork for delete project requests
Forking may lead to unbound processes running in the background,
using up all mem and killing the link info database. So don't do
it. Instead, send the answer early and just continue in the
same process. This will block the server slot, but that has the
wanted side-effect of limiting the delete projects that can run
in parallel.