[dist] V2: Use localhost instead of $HOSTNAME by default for obs worker
The ${HOSTNAME} variable of obsworker  service daemon is dependent
on the /etc/hostname file. Sometimes, we have often got the abnormal
execution of obsworker service daemon because of the incorrect mapping
situation of ${HOSTNAME}. In order to keep the stable execution and more
portability of the obswork service, Let's use localhost instead of the
/etc/hostname by default such as the "localhost:{5352|5252}" of
repserver/srcserver.

Ver2:
- Added $OBS_USE_MULTIPLE_WORKERS in /etc/obs-server file to support
an unique worker name in case that system has multiple obs workers.

Ver1:
- Use localhost instead of the $HOSTNAME by default in order to avoid
an abnormal execution of obsworker service daemon.

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>