[webui][ci] Remove render :text deprecation warnings
Running all rspec tests results in 50 lines of this kind at the log/tests.log file:

`render :text` is deprecated because it does not actually render a `text/plain` response. Switch to `render plain: 'plain text'` to render as `text/plain`, `render html: '<strong>HTML</strong>'` to render as `text/html`, or `render body: 'raw'` to match the deprecated behavior and render with the default Content-Type, which is `text/plain`.