[api][webui] Add validation for Comment type
The column `type` of Comment is reserved for storing the class in case
of inheritance and it is `nil` otherwise. As we validate that it is
present, it can not be `nil`. So that means that it is not possible to
create a Comment directly, only instances of Comment children can be
created. In case that we try to create an Comment directly with a non
`nil` type which is not a name of any of the Comment children an
ActiveRecord::SubclassNotFound is raised in the database level.

It is not obvious that an instance of Comment can not be directly
created and that the type is used for the single table inheritance.
It is also a good idea to check it in a Rails validation.