[backend] add batcher support to BSSrcdiff.pm
If we need to do lots of diffs for a bif tar archive, we call the diff
program from an extra process. While this sounds like something that
makes no sense at all, it actually speeds up diffing by a factor of 20.
The reason is that fork/exec is much slower if the process has a big
memory size.

A better approach would be a perl module that does the diffing with
no exec at all but alas, there seems to be no such thing (execpt a
pure perl version that is too slow).