[backend] Channel: always strip last package name part for maintenance incidents
We used to check the package links here. But the code in the API
was changed and the scheduler code was not adapted. Plus, we don't
have the link information in some cases.

Now we do it the easy way by just tripping of the last component
of the package name for incidents. In the future, we'll use a
new "releasetarget package" element from the package meta data.