Initial support for systemd unit files
This are the systemd unit files that we are using in open-build-service
Debian package. Most of the service units behave similar to the sysvinit
scripts, however it is to note that scheduler service is a templated
service unit to be able to launch an instance per architecture.

Note that logging facility changes to use systemd journal and drop the
wrapper support for logging in the OBSDIR/log directory.

Also note that not all services have been ported to systemd, so there are
a few services missing such obsworker.

In this patch, the systemd files are not being used in the packaging scripts.
Those would need to be enabled if someone wants to switch to them.

Signed-off-by: Héctor Orón Martínez <hector.oron@collabora.co.uk>