[api] Get rid of the nobody user thread caching
We are catching the nobody user in the User#find_nobody method and it is
only used here. It is confusing and  the Rails query cache will cache
this too, so we will only do one more query per thread.

Also, the https://github.com/openSUSE/open-build-service/pull/2247 PR is
failing because of this way of caching users and how the database is
cleaned, and it was difficult to get to know what the error was.