[api] dynamic :action segment is deprecated
Using a dynamic :action segment in a route is deprecated and will be
removed in Rails 5.1. I have changed

`get 'public/status/:action' => 'status#index'`

by

`get "public/status/method_name" => "status#method_name"`

for any of the method in the status controller.

Note that this is not exactly the same as what we used to have, as
before routes like `public/status/permitted_attributes` and
`public/status/auth_method` were available. But I think that we do not
want this routes to be available, so this fix the deprecation warning and
the fact that this kind of routes are available.