[api] Migration to use the old comments
Add a migration which moves all the comment's data to use the relatation
instroduced in e34e4156b628c6f1c0307b961da9fa290bdf99c6 That allows to
use the old comments.

Note that comments are only save if all the data can me moved. If not
the migration will fail.