[webui] Further improve speed on user/show page (related to issue #2201)
Limit the amount of records to fetch from DB. We only show a limited numbers
in the table, 25 by default. Thus we can limit the SQL query as well.

This improved the page load time from 16 to 22 seconds down to 5 to 8 seconds.

The test had to be adopted since the obsject_ids where causing a mismatch. Probably
caused by the changed scoping (using limit).