[api][webui] fix breakage on project destroy
Avoid the double write of meta data. It leads to a broken project
after undelete and it is causing double load in the backend.

Testcase for correct undelete behaviour added.

method in project model which was only used in one place got dropped,
since it wasn't usable this way anyway.