[api] Change Project & Package name type to String
There is a validation for both Package and Project that limit their
names to 200 characters, so we should use string of 200 characters
instead of text in the database. There is also a validation to avoid
that the name is nil, so the field should be not nullable in the
database.