[backend] remove duplicate indexes
Remove duplicate indexes as suggested by pt-duplicate-key-checker.

A key or unique constraint on "a,b" is also an index on just "a".
Such duplicate indexes accumulate over the lifetime of a sw project
and do not normally hurt in production. However keeping them around
carries a performance penalty as all indexes need to be maintained
by the database engines.

If a previous database migration class was skipped or structure.sql
does not represent the current schema, this potentially removes
useful indexes - this would, however, be an operational problem.

* src/api/db/migrate/20170123115500_remove_duplicate_indexes.rb
  Migration class
* structure.sql: changed my same DDL