[webui] extend the temporary hack of with an extra protection
if this wents wrong (because some implicit save via some depending
objects) we will only notice later due to corrupted projects, but we
won't have anything in the history of these projects. Means it will be
really hard to debug and to find. Given that this would be also a
grave security issue which gives me sleepless night a protection for
that seems to be a good idea.

Can go away again when the "FIXME" issue in webui package controller get solved.