[webui] give the user a hint about hidden build results
Can happen in case of infra structure problems, bugs, stopped schedulers,
lost events or administrator config changes for example. A remaining build
result may still have an effect to others builds and publishing (esp. in cross
architecture cases). We need to give the user at least a hint that the data
is not complete or he would get leaded now to look in a wrong direction.

FIXME: @project.repository.any? with build results case is not handled