[ci] Test source_controller#show_project_meta
It is tested in `project_controller_tests.rb#test_add_and_modify_repo` of the old test suite so there was a coverage difference between both test suites. I tested it with a simple project, instead with a complex one as it is done in the old test suite as it is not needed. It would be better to test the `Project#to_axml` method with a complex project than doing it here.