[webui] Modify the dialog.scss file for correctly displaying input boxes
If this would not be added the input box would have a length of 90%.
This would display the input field wrongly, like this:

https://cloud.githubusercontent.com/assets/8287131/23656314/82c34040-0339-11e7-8825-3b8bc6275fa8.png

This happens because the dialog does no longer gets loaded in
the #dialog_wrapper div. The old didn't get used during this
implementation, because dialogs, that gets loaded via Ajax, gets removed
from the DOM afterwards. If we would still use this, the dialog would
get removed from the DOM afterwards (that we don't want)