[api][ci] Validate updating package activity
`package#update_activity` was called from both a `before_update` and a method.
An attribute was updated inside this method and as this was call from a
`before update` could not be updated because if not it would enter in an
infinite loop. So `update_column`, which skip the validations, was used for
that. But then when being called from the method the validations were skip.

Now, instead of calling it from the mehotd we call `update` there, which calls
the `before_update` callback and run all the validations as well.