[api] don't allow to create requests with same target actions
It makes usually no sense to manipulate the same target with the same operation.
The exception is "add_role" since you can requests to add multiple
roles to the same target. And "maintenance_incident" since it does not
specify the target package

We saw this kind of error happening multiple times when having broken source
setup. It can lead to all kind of follow up errors, so let's protect us and
tell the user his problem as early as possible

Avoids/Hides also another bug leading to data corruption on release of such
a broken request