[api] Solve `password_changes_with_hash_type` bug
When the hash type is changed, the password should be updated, except if it was
`nil?`. In that case we don't care if it is updated or not. But that was not
what the method was doing. It failed when the password was not changed and and
when it was changed but it was `nil` before, as `!password_changed` was
`false` but `password_was.nil?` was `true`.