[ci] Autogenerate `.haml-lint_todo.yml`
The `auto-gen-config` option for `haml-lint` is not yet in the last release as
it was added to the project yesterday. Moreover, there are some bugs on the
current version of the gem in master. So, I used the version in:

https://github.com/Ana06/haml-lint/tree/fix-todo

to autogenerate the file.

I also added a comment in the generated file to point to the branch where the
code used to generate the file is.